reduced the number of hacks
now the `DragArea`s are dragged normally (more or less)
the only hack is that they are reparented as sibling of the event item
(i.e. the item being resized) to avoid incurring into the problem of a
`MouseArea` which is a child of the item being moved/resized.
Then the event item manually applies the changes in event time/duration
based on the position of the `DragAreas`s.

Also all the `MouseArea`s cursorShape property is still bound to the active
`DragArea`'s cursorShape to avoid cursorShape flickering.